server/requirements.txt
flask
flask-cors
flask-socketio
eventlet
werkzeug
pyjwt
